\name{bayeslr-package}
\alias{bayeslr-package}
\alias{bayeslr}
\docType{package}
\title{
\packageTitle{bayeslr}
}
\description{
\packageDescription{bayeslr}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{bayeslr}

Maintainer: \packageMaintainer{bayeslr}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = bayeslr)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{bayeslr}}
%\section{Documentation Index}{\packageIndices{bayeslr}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
