\docType{package}
\name{bayeslr}
\alias{bayeslr}
\alias{bayeslr-package}
\title{Functions for fitting and using Bayesian linear regression models.}
\usage{
bayeslr(w_mean, w_cov, sigsq)
}
\description{
Functions for fitting and using Bayesian linear regression
models.

'bayeslr' wraps up the model hyperparameters into a single
object that can later be used for inference.
}

